refactor(design-os): split evolution_core.py under the 200-line guideline (spec 012 P1)
heartbeat/registry/roles/taste-votes readers move to evolution_signals.py; evolution_core
re-imports and re-exports them so gather_signals' composition is unchanged. Adds the P1
report with both live verdicts (dana DEAD-LOOP, VSF-PCP ALIVE) and findings.
